+++ Assets/CommonCoreGame/RPGGame/World/Actor/ActorController.cs
@@ -300,7 +300,7 @@ namespace CommonCore.RpgGame.World
                             this.gameObject.SetActive(false); //actually destroying the object breaks saving
 
                         if (OnDeathSpecial != null)
-                            OnDeathSpecial.Execute(new ActionInvokerData { Activator = this });
+                            OnDeathSpecial.Execute(new ActionInvokerData { Activator = LastHit?.Originator, Caller = this, Position = transform.position, Rotation = transform.rotation, Velocity = MovementComponent.Ref()?.PhysicsVelocity });
 
                         if (DisableHitboxesOnDeath)
                         {